change migration for cards table

+++ db/migrate/20151006202215_drop_table_cards.rb
@@ -1,5 +1,6 @@
 class DropTableCards < ActiveRecord::Migration
   def change
-    drop_table :cards
+    if ActiveRecord::Base.connection.table_exists? 'cards'
+      drop_table :cards
   end
 end
